+++ src/server/express/a2a_express_app.ts
@@ -1,18 +1,16 @@
-import express, { Request, Response, Express, RequestHandler, ErrorRequestHandler, NextFunction } from 'express';
+import express, { Express, RequestHandler, ErrorRequestHandler } from 'express';
 
-import { A2AError } from "../error.js";
-import { JSONRPCErrorResponse, JSONRPCSuccessResponse, JSONRPCResponse } from "../../index.js";
 import { A2ARequestHandler } from "../request_handler/a2a_request_handler.js";
 import { JsonRpcTransportHandler } from "../transports/jsonrpc_transport_handler.js";
 import { AGENT_CARD_PATH } from "../../constants.js";
+import { jsonErrorHandler, jsonRpcHandler } from './json_rpc_handler.js';
+import { agentCardHandler } from './agent_card_handler.js';
 
 export class A2AExpressApp {
-    private requestHandler: A2ARequestHandler; // Kept for getAgentCard
-    private jsonRpcTransportHandler: JsonRpcTransportHandler;
+    private requestHandler: A2ARequestHandler;
 
     constructor(requestHandler: A2ARequestHandler) {
-        this.requestHandler = requestHandler; // DefaultRequestHandler instance
-        this.jsonRpcTransportHandler = new JsonRpcTransportHandler(requestHandler);
+        this.requestHandler = requestHandler;
     }
 
     /**
@@ -30,97 +28,21 @@ export class A2AExpressApp {
         agentCardPath: string = AGENT_CARD_PATH
     ): Express {
         const router = express.Router();
-        router.use(express.json(), ...(middlewares ?? []));
 
-        router.use((err: any, req: Request, res: Response, next: NextFunction) => {
-            // Handle JSON parse errors from express.json() (https://github.com/expressjs/body-parser/issues/122)
-            if (err instanceof SyntaxError && 'body' in err) {
-                const a2aError = A2AError.parseError('Invalid JSON payload.');
-                const errorResponse: JSONRPCErrorResponse = {
-                                jsonrpc: '2.0',
-                                id: null,
-                                error: a2aError.toJSONRPCError(),
-                            };
-                return res.status(400).json(errorResponse);
-            }
-            next(err);
-        });
+        // Doing it here to maintain previous behaviour of invoking provided middlewares
+        // after JSON body is parsed, jsonRpcHandler registers JSON parsing on the local router.
+        // body-parser used by express.json() ignores subsequent calls and is safe to be added twice:
+        // https://github.com/expressjs/body-parser/blob/168afff3470302aa28050a8ae6681fa1fdaf71a2/lib/read.js#L41.
+        router.use(express.json(), jsonErrorHandler);
 
-        router.get(`/${agentCardPath}`, async (req: Request, res: Response) => {
-            try {
-                // getAgentCard is on A2ARequestHandler, which DefaultRequestHandler implements
-                const agentCard = await this.requestHandler.getAgentCard();
-                res.json(agentCard);
-            } catch (error: any) {
-                console.error("Error fetching agent card:", error);
-                res.status(500).json({ error: "Failed to retrieve agent card" });
-            }
-        });
+        if (middlewares && middlewares.length > 0) {
+            router.use(middlewares);
+        }
 
-        router.post("/", async (req: Request, res: Response) => {
-            try {
-                const rpcResponseOrStream = await this.jsonRpcTransportHandler.handle(req.body);
-
-                // Check if it's an AsyncGenerator (stream)
-                if (typeof (rpcResponseOrStream as any)?.[Symbol.asyncIterator] === 'function') {
-                    const stream = rpcResponseOrStream as AsyncGenerator<JSONRPCSuccessResponse, void, undefined>;
-
-                    res.setHeader('Content-Type', 'text/event-stream');
-                    res.setHeader('Cache-Control', 'no-cache');
-                    res.setHeader('Connection', 'keep-alive');
-                    res.flushHeaders();
-
-                    try {
-                        for await (const event of stream) {
-                            // Each event from the stream is already a JSONRPCResult
-                            res.write(`id: ${new Date().getTime()}\n`);
-                            res.write(`data: ${JSON.stringify(event)}\n\n`);
-                        }
-                    } catch (streamError: any) {
-                        console.error(`Error during SSE streaming (request ${req.body?.id}):`, streamError);
-                        // If the stream itself throws an error, send a final JSONRPCErrorResponse
-                        const a2aError = streamError instanceof A2AError ? streamError : A2AError.internalError(streamError.message || 'Streaming error.');
-                        const errorResponse: JSONRPCErrorResponse = {
-                            jsonrpc: '2.0',
-                            id: req.body?.id || null, // Use original request ID if available
-                            error: a2aError.toJSONRPCError(),
-                        };
-                        if (!res.headersSent) { // Should not happen if flushHeaders worked
-                            res.status(500).json(errorResponse); // Should be JSON, not SSE here
-                        } else {
-                            // Try to send as last SSE event if possible, though client might have disconnected
-                            res.write(`id: ${new Date().getTime()}\n`);
-                            res.write(`event: error\n`); // Custom event type for client-side handling
-                            res.write(`data: ${JSON.stringify(errorResponse)}\n\n`);
-                        }
-                    } finally {
-                        if (!res.writableEnded) {
-                            res.end();
-                        }
-                    }
-                } else { // Single JSON-RPC response
-                    const rpcResponse = rpcResponseOrStream as JSONRPCResponse;
-                    res.status(200).json(rpcResponse);
-                }
-            } catch (error: any) { // Catch errors from jsonRpcTransportHandler.handle itself (e.g., initial parse error)
-                console.error("Unhandled error in A2AExpressApp POST handler:", error);
-                const a2aError = error instanceof A2AError ? error : A2AError.internalError('General processing error.');
-                const errorResponse: JSONRPCErrorResponse = {
-                    jsonrpc: '2.0',
-                    id: req.body?.id || null,
-                    error: a2aError.toJSONRPCError(),
-                };
-                if (!res.headersSent) {
-                    res.status(500).json(errorResponse);
-                } else if (!res.writableEnded) {
-                    // If headers sent (likely during a stream attempt that failed early), try to end gracefully
-                    res.end();
-                }
-            }
-        });
+        router.use(jsonRpcHandler({ requestHandler: this.requestHandler }));
+        router.use(`/${agentCardPath}`, agentCardHandler({ agentCardProvider: this.requestHandler }));
 
         app.use(baseUrl, router);
-        // The separate /stream endpoint is no longer needed.
         return app;
     }
 }
